Make all 16 host channels usable
The DWC USB controller IP supports up to 16 host channels.
In BCM283x, all 16 channels are implemented. Make them all available.

While we are at it, fix the mess of hardcoded channel counts (8 in some
places, others 16), and use a single define, NUM_CHANNELS, to simplify
changing this value in the future.