Change channel interrupt handling to match FreeBSD driver
Looks like the DWC core has some bugs related to signalling channel halt
interrupts in DMA mode. Rarely, it will signal the reason for a channel
halt, but not the fact itself that a channel has been halted. As a
workaround, interrupt on and check both the generic "channel halted" bit
and the individual reason bits. This is also what the FreeBSD driver
does (Linux is more complicated, as it uses FIQ which is not supported
in Windows.)